+++ Plugin.cs
@@ -15,7 +15,7 @@ public class Plugin : BasePlugin
 {
     public const string Guid = "bigorb";
     public const string Name = "Big Orb";
-    public const string Version = "1.1.0";
+    public const string Version = "1.1.1";
 
     internal static ManualLogSource Logger;
     internal static ConfigEntry<int> Port;